Add install scripts for doom-emacs and nvim. Currently only supports ubuntu

+++ nvim/install.sh
@@ -1,17 +1,24 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
-echo "Installing dependencies"
-sudo apt-get -y install nodejs npm
-sudo sudo npm install -g yarn@latest
+set -e
+cd $(dirname "$0")
 
-echo "Installing latest version of neovim"
-sudo apt remove neovim -y && \
-sudo add-apt-repository ppa:neovim-ppa/stable  && \
-sudo apt-get update && \
-sudo apt-get install neovim
+if [ -f /etc/lsb-release ]; then
+  echo "Running Neovim installation for Debian based system"
 
-echo "Installing config"
-mkdir ~/.config/nvim
-cp ,/init.vim ~/.config/nvim
+  echo "Installing dependencies"
+  sudo apt-get -y update
+  sudo apt-get -y install nodejs npm
+  sudo npm install -g yarn@latest
+
+  echo "Installing latest version of neovim"
+  sudo add-apt-repository ppa:neovim-ppa/stable -y
+  sudo apt-get update
+  sudo apt-get install neovim
+
+  echo "Installing config"
+  mkdir -p ~/.config/nvim
+  cp ./init.vim ~/.config/nvim
+fi
 
 echo "Done"
